requirements.txt
affine==2.3.1
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Django==4.0.3
Fiona==1.8.21
fonttools==4.31.2
GDAL==3.3.2
GeoAlchemy==0.7.2
GeoAlchemy2==0.10.0
geopandas==0.10.2
geoserver-rest==2.1.4
greenlet==1.1.2
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
munch==2.5.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pkg_resources==0.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pygdal==3.3.2.10
Pygments==2.11.2
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2022.1
rasterio==1.2.10
requests==2.27.1
Rtree==0.9.7
scipy==1.8.0
seaborn==0.11.2
Shapely==1.8.1.post1
six==1.16.0
snuggs==1.4.7
SQLAlchemy==1.4.32
sqlparse==0.4.2
urllib3==1.26.9
xmltodict==0.12.0
